feature: New restaurants activity added

+++ Dev/app/src/main/java/com/avielniego/openhvr/data/loadData/RestaurantCursorParser.java
@@ -64,7 +64,7 @@ public class RestaurantCursorParser {
     public RestaurantContent parse(Cursor data)
     {
         RestaurantContent content = new RestaurantContent();
-        content.id = data.getLong(RESTAURANT_ID_COLUMN_INDEX);
+        content.id = data.getInt(RESTAURANT_ID_COLUMN_INDEX);
         content.image = data.getString(COLUMN_RESTAURANT_IMAGE);
         content.name = data.getString(RESTAURANT_NAME_COLUMN_INDEX);
         content.desc = data.getString(COLUMN_RESTAURANT_DESC);
