+++ Dev/app/src/main/java/com/avielniego/openhvr/data/storedData/RestaurantContract.java
@@ -39,7 +39,7 @@ public class RestaurantContract
         public static final String CONTENT_TYPE      = ContentResolver.CURSOR_DIR_BASE_TYPE + "/" + CONTENT_AUTHORITY + "/" + PATH_RESTAURANT;
         public static final String CONTENT_ITEM_TYPE = ContentResolver.CURSOR_ITEM_BASE_TYPE + "/" + CONTENT_AUTHORITY + "/" + PATH_RESTAURANT;
 
-        public static Uri buildRestaurantUri(long id)
+        public static Uri buildRestaurantUri(int id)
         {
             return ContentUris.withAppendedId(CONTENT_URI, id);
         }
